refactor: extract getDelta duplication into recentLogsUtils (#196)
- Add src/lib/utils/recentLogsUtils.ts with computeWeightDelta / buildRecentLogArrays
- Remove identical getDelta / sorted / ascending boilerplate from RecentLogsTable and RecentLogsCards
- Both components now import shared helpers; behavior unchanged
